restructure repo and update notebooks
